requirements.txt
importlib-metadata~=1.0
python-dateutil~=2.0
web3~=5.0
pygit2~=1.0
docker~=4.0
